pyproject.toml
[project]
name = "first-fastapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
